frontend/requirements.txt
streamlit
requests-aws4auth
pandas
wikipedia
langchain
boto3